[FIX] account: prevent loops when unreconciling entries
df8fc9b2013f48ffc4c7d12b38756f5ba9e861d7 refactored the handling of
exchange rate entries when unreconciling entries.

However it did not preserve the piece of code that prevented loops
during this operation (`full_rec_lookup` context key).

This could cause the unreconciliation of entries with an exchange rate
difference to fail with a confusing error claiming that the user is
trying "to reconcile some entries that are already reconciled!".
The double reconciliation was caused by a recursive attempt to reverse
the exchange rate entry.