[FIX] web: allow edition of one2many with widget many2many
It is quite rare, but it happens that a one2many field has to be edited
with a many2many widget: it allows the fact that records are directly
created, and only later linked to the main record.  But the BasicModel
only used the field type to determine which commands it had to send.

In this commit, we introduce a magic flag _forceM2MLink, which, if set
to true, tells the BasicModel to treat the field as a many2many.