[FIX] web: BasicModel: recursively fetch x2manys
Before this commit, x2many fields embedded in x2many fields were
not fetched. This was for example the case for many2many_tags in
one2many fields.

This commit enables the recursive fetch of x2many fields.

There is still a missing piece, as the list of fields to fetch for
those embedded x2manys is still unknown. This will come in a
future commit.