[IMP] mail: add activity types configuration that is missing to use activities efficiently
Currently tweaking or defining new activity types is difficult to achieve
as menu is hidden under technical settings. This commit eases the activity
types configuration by adding a menu allowing to customize them directly
under Emails section of General Settings. By the way it is renamed to
Discuss to better reflect its purpose.

This is an improvement but considered as an usability fix, therefore
targeting a stable version.