[FIX] stock: multi-step routes
When we are changing the incoming and outgoing routes steps of a
warhouse through its settings, the push/pull rules are actually deleted
and recreated, which raise a traceback if we have some move validated
that are using those rules.

The problem came after rev: https://github.com/odoo/odoo/commit/11dc19e9f5bb0a57772f61da70e49f3aa7a7cef6
Because we restrict the deletion of rules when they are used in stock
moves.

To fix this issue, we deactivate rules when we change the incoming and
outgouing rules instead of unlink them