[IMP] hr_timesheet, sale_timesheet: log timesheet on employee
This commits mainly switches the semantic of timesheeting from users
to employees. Main points of this commit are :

 * add an employee_id field on timesheet;
 * replace user_id field by employee_id on timesheet views;
 * provide a stat button on employee form view to access its timesheets;
 * use timesheet cost as a real monetary field and handle currency
   conversion;

The purpose is to allow timesheet manager to record timesheet for
employees that are not users. All employees do not necessarily have an
user account in the system. Now you can record timesheet for your employee.
Hooray !

In sale_timesheet timesheet_cost is now a monetary field. Currency convertion
is also better taken into account when computing UPDATE ME

Timesheet details

 * add an invisible timesheet group on employee form view allowing to
   hook on it. Indeed sale_timesheet and timesheet_grid modules override
   this view to add information but have no other common module that
   hr_timesheet. Using an invisible group set as visible in those modules
   eases inheritance.

Sale timesheet details

 * timesheet_cost is now a monetary field
 * better display of timesheet_cost on employee form view, notably telling
   it is per hour
 * remove account_id field on employee as its only use was buggy and was
   mixing account.account and analytic.account records.