[FIX] web: BasicModel: list count when grouped
Before this rev., the count attribute of list dataPoints was
basically the length of the data Array. It means that for an
ungrouped list, it was the number of records in the list, and for
a list containing groups, it was the number of groups.

This caused a problem in the list view with nested groups, as the
counter in the group header is supposed to display the number of
records in the group, whether or not the group contains other
subgroups or directly records.

This commit changes a bit the semantics of the count attribute,
such that it always represents the number of records inside the
list (i.e. the sums of records inside subgroups if the list is
grouped).