[IMP] rating: parenting model mecanism
This commit provide a parenting system for rated
record, such as mail alias.
For instance, 'project.project' is the parent of
task and issue objects.

This will allow to have a global view of customer
statisfaction on 'parent' objects (project and
livechat channel for now).

To do so, the rating
mixin provides 2 methods that need to be
overriden: `rating_get_parent_model_name` and
`rating_get_parent_id`.